Put mnv3 impls to one file
